requirements.txt
aiogram
python-dotenv
yt-dlp
pytest
redis